Refactor board module to separate folder and update class name
